removing warnings/errors from sdk
- it is not very useful to have -Werror for the vst3 sdk itself, especially since with Xcode 14+, many warnings trigger an error thus failing the build entirely

+++ src/cpp/pongasoft/VST/Messaging.h
@@ -157,7 +157,7 @@ tresult Message::setSerializableValue(IAttributeList::AttrID id, const IParamSer
   tresult res = iSerializer.writeToStream(iValue, streamer);
   if(res == kResultOk)
   {
-    return setBinary(id, stream.getData(), stream.getSize());
+    return setBinary(id, stream.getData(), static_cast<uint32>(stream.getSize()));
   }
   return res;
 }
